fix: change CartItem id type

+++ Controllers/ProductController.cs
@@ -41,7 +41,7 @@ namespace TallerIDWMBackend.Controllers
 
         // POST: api/product/add-to-cart/{id}
         [HttpPost("add-to-cart/{id}")]
-        public async Task<IActionResult> AddToCart(int id)
+        public async Task<IActionResult> AddToCart(long id)
         {
             var product = await _productRepository.GetProductByIdAsync(id);
 
@@ -52,7 +52,7 @@ namespace TallerIDWMBackend.Controllers
 
             // Gestionar el carrito de compras con cookies
             var cart = Request.Cookies["cart"];
-            var cartItems = string.IsNullOrEmpty(cart) ? new List<int>() : cart.Split(",").Select(int.Parse).ToList();
+            var cartItems = string.IsNullOrEmpty(cart) ? new List<long>() : cart.Split(",").Select(long.Parse).ToList();
 
             // Verificar si el producto ya está en el carrito
             if (cartItems.Contains(id))
